requirements.txt
# Librerías básicas
pandas==2.0.3
numpy==1.24.3

# Base de datos
pymysql==1.1.0
sqlalchemy==2.0.20
python-dotenv==1.0.0

# Visualización
matplotlib==3.7.2
seaborn==0.12.2

# Machine Learning
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.1.0
imbalanced-learn==0.11.0

# Jupyter
jupyter==1.0.0
ipykernel==6.25.1
